+++ src/proxy/domainmap.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012-2020 Fanout, Inc.
+ * Copyright (C) 2012-2022 Fanout, Inc.
  *
  * This file is part of Pushpin.
  *
@@ -134,8 +134,8 @@ public:
 	class Entry
 	{
 	public:
-		QByteArray pathBeg;
 		QByteArray id;
+		QByteArray pathBeg;
 		QByteArray sigIss;
 		QByteArray sigKey;
 		QByteArray prefix;
@@ -150,6 +150,8 @@ public:
 		QByteArray sockJsPath;
 		QByteArray sockJsAsPath;
 		HttpHeaders headers;
+		bool separateStats;
+		bool grip;
 		QList<Target> targets;
 
 		bool isNull() const
@@ -157,12 +159,22 @@ public:
 			return targets.isEmpty();
 		}
 
+		QByteArray statsRoute() const
+		{
+			if(separateStats)
+				return id;
+			else
+				return QByteArray(); // global stats
+		}
+
 		Entry() :
 			origHeaders(false),
 			pathRemove(0),
 			debug(false),
 			autoCrossOrigin(false),
-			session(false)
+			session(false),
+			separateStats(false),
+			grip(true)
 		{
 		}
 	};
@@ -175,7 +187,9 @@ public:
 	//   underlying file watching doesn't work
 	void reload();
 
+	bool isIdShared(const QString &id) const;
 	Entry entry(Protocol proto, bool ssl, const QString &domain, const QByteArray &path) const;
+	Entry entry(const QString &id) const;
 
 	QList<ZhttpRoute> zhttpRoutes() const;